+++ internal/pkg/diff/diff.go
@@ -10,9 +10,9 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
-	"github.com/keboola/keboola-as-code/internal/pkg/strhelper"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/orderedmap"
+	"github.com/keboola/keboola-as-code/internal/pkg/utils/strhelper"
 )
 
 type typeName string
