+++ internal/pkg/utils/strhelper/strhelper.go
