+++ internal/pkg/utils/strhelper/strhelper_test.go
